Rerank endpoint parameters and max document check (#2814)
CVS-156600
