+++ src/rerank/rerank_utils.cpp
@@ -194,7 +194,7 @@ absl::Status chunkDocuments(
     const ov::Tensor& in_input_ids, const ov::Tensor& in_attention_mask,
     ov::Tensor& out_input_ids, ov::Tensor& out_attention_mask,
     std::vector<size_t>& chunk_mapping, size_t max_tokens_per_chunk,
-    int64_t pad_token) {
+    size_t max_allowed_chunks, int64_t pad_token) {
 
     if (max_tokens_per_chunk == 0) {
         return absl::InvalidArgumentError("no space left for chunks");
@@ -218,6 +218,9 @@ absl::Status chunkDocuments(
 
     size_t tokens_count_of_longest_document = in_input_ids.get_shape()[1];
     size_t batch_size = in_input_ids.get_shape()[0];
+    if (batch_size > max_allowed_chunks) {
+        return absl::InvalidArgumentError(std::string{"exceeding max_allowed_chunks before chunking limit: "} + std::to_string(max_allowed_chunks) + std::string{"; actual: "} + std::to_string(batch_size));
+    }
 
     if (tokens_count_of_longest_document <= max_tokens_per_chunk) {
         out_input_ids = in_input_ids;
@@ -245,6 +248,9 @@ absl::Status chunkDocuments(
     }
 
     size_t new_batch_size = chunk_mapping.size();
+    if (new_batch_size > max_allowed_chunks) {
+        return absl::InvalidArgumentError(std::string{"exceeding max_allowed_chunks after chunking limit: "} + std::to_string(max_allowed_chunks) + std::string{"; actual: "} + std::to_string(new_batch_size));
+    }
 
     if (new_batch_size != batch_size) {
         SPDLOG_LOGGER_DEBUG(rerank_calculator_logger, "Chunking required, initial batch size: {}, final batch size: {}", batch_size, new_batch_size);
