+++ src/rerank/rerank_utils.hpp
@@ -76,6 +76,6 @@ absl::Status chunkDocuments(
     const ov::Tensor& in_input_ids, const ov::Tensor& in_attention_mask,
     ov::Tensor& out_input_ids, ov::Tensor& out_attention_mask,
     std::vector<size_t>& chunk_mapping, size_t max_tokens_per_chunk,
-    int64_t pad_token);
+    size_t max_allowed_chunks, int64_t pad_token);
 
 }  // namespace ovms
